requirements.txt
streamlit
langchain
langchain_google_genai
langchain-community
faiss-cpu
Pillow
# psycopg2
pytesseract
# dotenv